feat: introduce versioned KV strategy
